### 2017-04-06 #109 @kaneplusplus
Submission for R Package ccex
Package ccex is an R implementation of the REST interface used by the C-Cex crypto-currency exchange. It provides functions for endpoints supported by the exchange. This includes the ability to retrieve price, volume, and order book information as well as the ability to trade crypto-currencies.

```
Package: ccex
Title: Client for the C-Cex Crypto-currency Exchange
Version: 0.1.0
Authors@R: c(person("Michael", "kane", role = c("aut", "cph", "cre"), email = "kaneplusplus@gmail.com"))
Description: A client for the C-Cex crypo-currency exchange \url{https://c-cex.com} including the ability to query trade data, manage account balances, and place orders.
License: LGPL-2
Imports: httr, openssl, stringr
Suggests: covr, testthat, knitr, rmarkdown, threejs, tidyverse, igraph, quantmod
```

(https://github.com/kaneplusplus/ccex)[https://github.com/kaneplusplus/ccex]

Target audience is finance researchers and traders interested in crypo-currencies.

No package provide an interface to the (C-Cex cryptocurrency exchange)[https://c-cex.com/] meant for research, which includes the ability to trade. Package (Rbitcoin)[https://github.com/jangorecki/Rbitcoin] is an "end-to-end trading engine in R" for currency trading with Bitcoin. It is not clear of other crypto-currencies are included. Package (IBroker)[https://github.com/joshuaulrich/IBrokers] is an API for the (Interactive Broker)[https://www.interactivebrokers.com/en/home.php] trading platform. It provides the ability to trade stocks and currencies, not crypto-currencies.

The `devtools::check()` command succeeds with no warnings.

No exceptions

No.

Joshua Ulrich: joshuaulrich
Whit Armstrong: armstrtw
