### 2017-04-10 90afac2
Removed dependency on stringr.
