+++ NAMESPACE
@@ -23,5 +23,3 @@ importFrom(httr,GET)
 importFrom(httr,add_headers)
 importFrom(httr,content)
 importFrom(openssl,sha512)
-importFrom(stringr,str_replace_all)
-importFrom(stringr,str_sub)
