### 2017-04-11 #109 @noamross
Thank you for your submission, @kaneplusplus! You may be interested to know that an early rOpenSci contribution was https://github.com/ropensci/RCryptsy, though that service has since closed and the package removed from CRAN.

### Editor checks:

---

#### Editor comments

I noted I needed the github/development version of **threejs** to build the vignette.  You may wish to add this to `Remotes:` in your DESCRIPTION file, but in any case you'll want a CRAN-stable version if you are submitting to CRAN.

Here is output from `goodpractice::gp()`. As the package passes R CMD CHECK with no errors, warnings, or notes, these won't hold up assigning reviewers. Please address them during the review, though.  100% coverage isn't required but reviewers can use this to examine whether tests cover adequate use-cases.

```
── GP ccex ─────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 86%
    of code lines are covered by test cases.

    R/ccex.r:583:NA
    R/ccex.r:584:NA
    R/ccex.r:585:NA
    R/ccex.r:588:NA
    R/ccex.r:589:NA
    ... and 16 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite
    often. A build date will be added to the package when you perform `R CMD
    build` on it.
  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users
    and developers are used it and it is easier to read your code for them if
    you use '<-'.

    R/ccex.r:61:15
    R/ccex.r:65:16
    R/ccex.r:95:7
    R/ccex.r:119:8
    R/ccex.r:120:8
    ... and 128 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R/ccex.r:240:1
    R/ccex.r:318:1
    R/ccex.r:342:1
    R/ccex.r:391:1
    R/ccex.r:432:1
    ... and 2 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...)
    expressions. They are error prone and result 1:0 if the expression on the
    right hand side is zero. Use seq_len() or seq_along() instead.

    R/ccex.r:283:12
    R/ccex.r:284:15
    R/ccex.r:583:14
    R/ccex.r:584:17
    R/ccex.r:628:14
    ... and 3 more lines

──────────────────────────────────────────
```
---

Reviewers:
Due date:
