+++ DESCRIPTION
@@ -8,6 +8,7 @@ BugReports: https://github.com/kaneplusplus/ccex/issues
 Description: A client for the C-Cex crypo-currency exchange \url{https://c-cex.com} including the ability to query trade data, manage account balances, and place orders.
 License: LGPL-2
 Imports: httr, openssl
+Remotes: bwlewis/rthreejs
 Suggests: covr, testthat, knitr, rmarkdown, threejs, tidyverse, igraph, quantmod
 VignetteBuilder: knitr
 RoxygenNote: 6.0.1
