### 2017-04-20 #109 @noamross
Reviewers: @statsmaths @kylehamilton
Due date: 2017-05-11
