### 2017-05-01 kaneplusplus/ccex#1 @taylor-arnold
ccex exchange detecting DDoS
I just tried running the tests and vignettes in the package and was hit with a large number of errors. It seems that there is some sort of problem with the web API:
```
$ curl -I https://c-cex.com/t/pairs.json
HTTP/1.1 503 Service Temporarily Unavailable
Date: Mon, 01 May 2017 14:21:42 GMT
Content-Type: text/html; charset=UTF-8
Connection: close
Set-Cookie: __cfduid=dcb62a0e73956a84b4364d69028f12cfe1493648502; expires=Tue, 01-May-18 14:21:42 GMT; path=/; domain=.c-cex.com; HttpOnly
X-Frame-Options: SAMEORIGIN
Refresh: 8;URL=/cdn-cgi/l/chk_jschl?pass=1493648506.344-lM8MLGKWef
Cache-Control: no-cache
Server: cloudflare-nginx
CF-RAY: 35836383a956690e-CDG
```
Grabbing the page through a browser works with a long delay where the warning "Checking your browser before accessing c-cex.com.".
