### 2017-05-01 #109 @statsmaths
I ran into issues with the package that seem to be related to a problem with the ccex API. Opened an issue at [https://github.com/kaneplusplus/ccex/issues/1](https://github.com/kaneplusplus/ccex/issues/1) and @kaneplusplus is looking into it.
