### 2017-05-01 kaneplusplus/ccex#1 @kaneplusplus
The following is the response I received from the exchange's support:

"Yes, it's temporary. We will remove limitations soon."

I'll keep track and send a note when it's removed.
