### 2017-05-02 #109 @noamross
Thanks for the update, @statsmaths and @kaneplusplus.
