### 2017-05-09 #109 @noamross
Hello, @kaneplusplus, I just want to double check on the status of this review. Are we waiting for the CCEX API to update before moving forward? If so, I'll pause the clock until you let us know so that @kylehamilton and @statsmaths can plan for when to do their reviews.
