### 2017-05-09 #109 @kaneplusplus
Hi @noamross. The C-cex exchange was DDos'ed and the exchange's cloud provider (CloudFlare) has limited traffic to the browser. The package and documentation are complete but the reviewers can't use the package to query the exchange until the limitation is lifted.

The exchanges support has verified that the limitation will be lifted "soon" but they have not given any idea how soon. I've been checking daily to see if it has been lifted.

I'm sorry about the inconvenience and I'm happy to handle this however you want.
