+++ R/ccex.r
@@ -13,7 +13,7 @@
 #' 
 #' Package 'ccex' is an R implementation of the REST interface used by the C-Cex
 #' crypto-currency exchange \url{https://c-cex.com/}. It provides functions 
-#' for all endpoints currently (as of April 1, 2017) supported by the 
+#' for all endpoints currently (as of May 16, 2017) supported by the 
 #' exchange. This includes the ability 
 #' to retrieve price, volume, and orderbook information as well as the ability
 #' to trade crypto-currencies.
