### 2017-05-19 047f559
Updated some documentation.
