### 2017-05-31 #109 @kaneplusplus
@noamross it looks like traffic is still being limited to C-cex. Could we have the same reviewers look at my [bittrex](https://github.com/kaneplusplus/bittrex) package instead? The difference is the [Bittrex crypto-currency exchange](https://bittrex.com) hasn't been limiting traffic for the last few weeks. The goal of the package is the same and the interface is almost the same.
