### 2017-06-01 #109 @noamross
Sorry that the API has been shut down in the middle of all this, @kaneplusplus.  Could you submut **bitrex** as a separate issue and I'll process it within a day or two?  @statsmaths and @kylehamilton, if so, would you be OK serving as reviewers for that?
