### 2017-06-06 kaneplusplus/ccex#1 @pothman
The same again. All API-calls rejected.
